tests/wpt/harness/test/testdata/testharness/testharness_0.html

<!doctype html>
<title>Simple testharness.js usage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_true(true);
}, "Test that should pass");

test(function() {
  assert_true(false);
}, "Test that should fail");
</script>
